x86: add support for Mellanox Spectrum Switches
The Mellanox Spectrum Switches are Managed Ethernet Switches
with speeds up to 800Gb/s. They utilize the switchdev API
to offload the kernel's forwarding plane to a real ASIC.

Tested on a Mellanox Spectrum SN2100 with the following specs:
 - CPU: Intel ATOM x86 dual-core 2.4GHz
 - RAM: 8GB
 - Disk: 16GB SSD
 - Ports: 16x QSFP28 100GbE, 1x 100M Mgmt Port, 1x RJ45 Serial Port
 - USB: 1x mini 2.0
 - Button: 1x (reset)
 - LEDs: 6x

Installation:
 - Create a bootable USB device (either by flashing this image
   onto it or another Linux distribution)
 - Unzip the image
   (openwrt-x86-64-mlnx-spectrum-ext4-combined-efi.img.gz)
   and copy it onto the USB device
 - Plug the USB device into the Mellanox Switch and boot from it
 - Flash the image (e.g., with dd) onto the internal SSD
   of the switch (should be /dev/sda)

To enter the BIOS, reboot the switch and press CTRL+B while you see
the BIOS information text (American Megatrends …). The default password
to enter the BIOS is 'admin'. To boot from the USB device, switch to the
'Save & Exit' tab and select your USB device from the Boot Override options.

Signed-off-by: Til Kaiser <mail@tk154.de>